workflow/requirements.txt
imap-tools>=1.7.0
openai>=1.40.0
httpx>=0.27.0
paramiko>=3.4.0
python-frontmatter>=1.1.0
PyYAML>=6.0.1
